Implement F instructions

+++ src/Emu/STRAIGHT64Linux/STRAIGHT64Operation.h
@@ -148,7 +148,7 @@ struct STRAIGHT64Addr : public std::unary_function<EmulatorUtility::OpEmulationS
 {
     STRAIGHT64RegisterType operator()(EmulatorUtility::OpEmulationState* opState) const
     {
-        return TSrc1()(opState) + EmulatorUtility::cast_to_signed(TSrc2()(opState));
+        return static_cast<u32>(TSrc1()(opState) + EmulatorUtility::cast_to_signed(TSrc2()(opState)));
     }
 };
